+++ packages/isaacscript-common/src/types/PossibleStatType.ts
@@ -1,0 +1,12 @@
+import { TearFlag } from "isaac-typescript-definitions";
+
+/**
+ * These are the possible types that player stats can be. For example, `EntityPlayer.Damage` is of
+ * type `float`, `EntityPlayer.CanFly` is of type `boolean`, and so on.
+ */
+export type PossibleStatType =
+  | number
+  | boolean
+  | BitFlags<TearFlag>
+  | Color
+  | Vector;